+++ lib/default.nix
@@ -48,8 +48,6 @@ let
       internalPercentDecode = callPackage ./internalPercentDecode.nix { };
     in
     {
-      craneLib = self;
-
       appendCrateRegistries = input: self.overrideScope (_final: prev: {
         crateRegistries = prev.crateRegistries // (lib.foldl (a: b: a // b) { } input);
       });
@@ -75,6 +73,7 @@ let
       cleanCargoToml = callPackage ./cleanCargoToml.nix { };
       configureCargoCommonVarsHook = callPackage ./setupHooks/configureCargoCommonVars.nix { };
       configureCargoVendoredDepsHook = callPackage ./setupHooks/configureCargoVendoredDeps.nix { };
+      craneLib = self;
       craneUtils = callPackage ../pkgs/crane-utils { };
 
       crateNameFromCargoToml = callPackage ./crateNameFromCargoToml.nix {