+++ lib/default.nix
@@ -129,8 +129,8 @@ let
       registryFromDownloadUrl = callPackage ./registryFromDownloadUrl.nix { };
       registryFromGitIndex = callPackage ./registryFromGitIndex.nix { };
       registryFromSparse = callPackage ./registryFromSparse.nix { };
-      removeReferencesToVendoredSourcesHook = callPackage ./setupHooks/removeReferencesToVendoredSources.nix { };
       removeReferencesToRustToolchainHook = callPackage ./setupHooks/removeReferencesToRustToolchain.nix { };
+      removeReferencesToVendoredSourcesHook = callPackage ./setupHooks/removeReferencesToVendoredSources.nix { };
       replaceCargoLockHook = callPackage ./setupHooks/replaceCargoLockHook.nix { };
       taploFmt = callPackage ./taploFmt.nix { };
       urlForCargoPackage = callPackage ./urlForCargoPackage.nix { };