+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+* `cargoNextest` now supports passing `cargoNextestPartitionsExtraArgs` to each
+  `cargo nextest` partition run.
+
 ## [0.20.0] - 2024-12-21
 
 ### Changed