+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+* Add self-reference `craneLib` to crane lib.
+
 ## [0.20.0] - 2024-12-21
 
 ### Changed