+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * `cargoNextest` now supports passing `cargoNextestPartitionsExtraArgs` to each
   `cargo nextest` partition run.
 * Add self-reference `craneLib` to crane lib instance.
+* Add `removeReferencesToRustToolchainHook`, which optionally scrubs the output 
+  binary of references to the Rust toolchain.
 
 ## [0.20.0] - 2024-12-21