+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [0.20.1] - 2025-02-08
+
 ### Added
 * `cargoNextest` now supports passing `cargoNextestPartitionsExtraArgs` to each
   `cargo nextest` partition run.
@@ -840,6 +842,7 @@ files parsed as nix attribute sets.
 ## 0.1.0 - 2022-01-22
 - First release
 
+[0.20.1]: https://github.com/ipetkov/crane/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/ipetkov/crane/compare/v0.19.4...v0.20.0
 [0.19.4]: https://github.com/ipetkov/crane/compare/v0.19.3...v0.19.4
 [0.19.3]: https://github.com/ipetkov/crane/compare/v0.19.2...v0.19.3